agent/pyproject.toml
[project]
name = "la-factoria"
version = "0.1.0"
description = "FastAPI agents template with React agent pattern, streaming API, and production-ready observability"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    { name = "Xtillion" }
]
keywords = ["fastapi", "agents", "react", "ai", "llm", "postgresql", "sqlalchemy", "streaming", "api"]
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.13",
    "Framework :: FastAPI",
]

dependencies = [
    # Core Application Dependencies
    "fastapi==0.119.0",
    "pydantic-settings==2.11.0",
    "uvicorn[standard]==0.37.0",
    "SQLAlchemy==2.0.44",
    "asyncpg==0.30.0",
    # Alembic Dependencies
    "psycopg2-binary==2.9.11",
    "alembic==1.17.0",
    # Logging Dependencies
    "structlog==25.4.0",
    "colorama==0.4.6",
    # OpenTelemetry Dependencies
    "opentelemetry-api==1.37.0",
    "opentelemetry-sdk==1.37.0",
    "opentelemetry-exporter-otlp==1.37.0",
    "opentelemetry-instrumentation-fastapi==0.58b0",
    "opentelemetry-instrumentation-sqlalchemy==0.58b0",
    "opentelemetry-instrumentation-asyncpg==0.58b0",
    # Additional Unorganized Dependencies
    "langgraph==1.0.0",
    "langchain==1.0.0",
    "langchain-openai==1.0.0",
    "langchain-anthropic>=1.0.3",
    "langchain-google-genai>=3.0.3",
    "langchain-aws>=1.0.0",
    "fastmcp>=2.13.1",
    "langchain-mcp-adapters>=0.1.14",
    "greenlet>=3.2.4",
    "langgraph-checkpoint-postgres>=3.0.2",
    "psycopg[binary,pool]>=3.3.2",
]

[project.optional-dependencies]
dev = [
    "black==25.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 120
target-version = ['py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | alembic
)/
'''
